# api/openapi/sbs-api-v1.yaml
openapi: 3.1.0
info:
  title: SBS SupTech Complaints API
  version: 0.2.0
  summary: Institution-to-regulator complaint reporting against Anexo 1-A.
  description: |
    The SBS SupTech Complaints API is the institution-facing surface through
    which supervised institutions (banks, financieras, AFPs, COOPACs,
    insurers, and other entities under SBS supervision) submit complaint
    records to the Superintendencia de Banca, Seguros y AFP (SBS) for
    regulatory analysis.

    Two ingestion tiers are offered. Tier 1 is a near-real-time per-complaint
    HTTP API for institutions with mature integration capability. Tier 2 is
    a signed batch upload for institutions whose operational capacity does
    not yet support per-event submission. Both tiers carry the same Anexo 1-A
    payload schema and the same validation rules.

    This v0.1.0 contract is the May 25, 2026 sprint kickoff scope. The
    payload model implements a curated 15-field subset of the Anexo 1-A
    taxonomy from Resolución SBS N° 04036-2022. Full Anexo 1-A coverage
    arrives in a later release alongside the Standards Pack v1.0.0
    distribution (see PLAN.md Part 11).

    See ADR 0026 for the field-by-field reconciliation of the 15-field
    subset against the resolución, and ADR 0027 for the rule that this
    specification — not the Pydantic implementation — is the canonical
    contract.

    ## complaint_id uniqueness

    The `complaint_id` is unique **globally within the SBS sandbox**, not
...

...

...
servers:
...
security:
  - mtls: []
    oauth2_client_credentials: [complaints:read]
    hmac_signature: []
paths:
  /complaints:
    post:
      tags: [Ingestion (Tier 1)]
      summary: Submit a single complaint
      operationId: createComplaint
      security:
        - mtls: []
          oauth2_client_credentials: [complaints:write]
          hmac_signature: []
      description: |
        Submit one complaint to SBS. The body carries the Anexo 1-A 15-field
        payload wrapped in a submission envelope. The request must carry an
        `Idempotency-Key` HTTP header (UUID v4 recommended) for at-least-once
        delivery semantics. On retry with the same key within 24 hours, the
        server replays the original response without re-processing.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - $ref: "#/components/parameters/TraceparentHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintSubmission"
            examples:
              valid:
                $ref: "#/components/examples/ValidComplaintSubmission"
      responses:
        "201":
          description: Complaint accepted.
          headers:
            traceparent: { $ref: "#/components/headers/TraceparentResponse" }
            Idempotency-Replayed: { $ref: "#/components/headers/IdempotencyReplayedResponse" }
            ETag: { $ref: "#/components/headers/ETagResponse" }
            Location:
              schema:
                type: string
                format: uri
              description: URL of the created complaint resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintCreated"
              examples:
                created:
                  $ref: "#/components/examples/ComplaintCreatedExample"
        "400": { $ref: "#/components/responses/Problem400" }
        "401": { $ref: "#/components/responses/Problem401" }
        "403": { $ref: "#/components/responses/Problem403" }
        "409": { $ref: "#/components/responses/Problem409" }
        "422": { $ref: "#/components/responses/Problem422" }
        "429": { $ref: "#/components/responses/Problem429" }
        "500": { $ref: "#/components/responses/Problem500" }
    get:
      tags: [Ingestion (Tier 1)]
      summary: List complaints
      operationId: listComplaints
      description: |
        Retrieve a paginated list of complaints. Supervisor users may filter
        across institutions; institution users see only their own. Pagination
        is cursor-based; treat `next_cursor` as **opaque** and pass it back
        verbatim. Internally the cursor encodes
        `{"received_at": "<RFC3339>", "complaint_id": "<uuid>"}` (base64),
        but this is not part of the published contract — see the info-block
        "Cursor pagination format". Inherits the global `complaints:read` scope.
      parameters:
        - in: query
          name: institution_id
          schema: { type: string, pattern: "^SBS-\\d{4,6}$" }
        - in: query
          name: received_date_from
          schema: { type: string, format: date }
        - in: query
          name: received_date_to
          schema: { type: string, format: date }
        - in: query
          name: product_category
          schema: { $ref: "#/components/schemas/ProductCategory" }
        - in: query
          name: channel
          schema: { $ref: "#/components/schemas/Channel" }
        - in: query
          name: motivo_code
          schema: { $ref: "#/components/schemas/MotivoCode" }
        - in: query
          name: resolution_status
          schema: { $ref: "#/components/schemas/ResolutionStatus" }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: next_cursor
          schema: { type: string, minLength: 8, maxLength: 256 }
        - $ref: "#/components/parameters/TraceparentHeader"
      responses:
        "200":
          description: Page of complaints.
          headers:
            traceparent: { $ref: "#/components/headers/TraceparentResponse" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintListResponse"
              examples:
                page:
                  $ref: "#/components/examples/ComplaintListExample"
        "400": { $ref: "#/components/responses/Problem400" }
        "401": { $ref: "#/components/responses/Problem401" }
        "403": { $ref: "#/components/responses/Problem403" }
        "422": { $ref: "#/components/responses/Problem422" }
        "429": { $ref: "#/components/responses/Problem429" }
        "500": { $ref: "#/components/responses/Problem500" }
  /complaints/{complaint_id}:
    parameters:
      - $ref: "#/components/parameters/ComplaintIdPath"
      - $ref: "#/components/parameters/TraceparentHeader"
    get:
      tags: [Ingestion (Tier 1)]
      summary: Retrieve a complaint by id
      operationId: getComplaintById
      description: Return the full Anexo 1-A complaint payload for a given complaint_id. Inherits the global `complaints:read` scope.
      responses:
        "200":
          description: Complaint found.
          headers:
            traceparent: { $ref: "#/components/headers/TraceparentResponse" }
            ETag: { $ref: "#/components/headers/ETagResponse" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
              examples:
                complaint:
                  $ref: "#/components/examples/ValidComplaintBody"
        "401": { $ref: "#/components/responses/Problem401" }
        "403": { $ref: "#/components/responses/Problem403" }
        "404": { $ref: "#/components/responses/Problem404" }
        "500": { $ref: "#/components/responses/Problem500" }
  /complaints/{complaint_id}/status:
    parameters:
      - $ref: "#/components/parameters/ComplaintIdPath"
      - $ref: "#/components/parameters/IdempotencyKeyHeader"
      - $ref: "#/components/parameters/IfMatchHeader"
      - $ref: "#/components/parameters/TraceparentHeader"
    patch:
      tags: [Ingestion (Tier 1)]
      summary: Update the resolution status of a complaint
      operationId: patchComplaintStatus
      security:
        - mtls: []
          oauth2_client_credentials: [complaints:write]
          hmac_signature: []
      description: |
        Transition a complaint's resolution status. A reason is required when
        transitioning to `atendido` or `anulado`. The audit log records both
        the prior and new state along with the reason. The response is the
        full updated `Complaint` resource so clients can observe the
        post-transition state without an extra round trip.

        Concurrency control: clients send the `ETag` returned from a prior
        `GET` or `PATCH` as the `If-Match` header. A stale tag returns 412
        with `ETAG_MISMATCH`; a missing tag returns 428.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintStatusPatch"
            examples:
              resolve:
                summary: Resolve a complaint
                value:
                  resolution_status: atendido
                  reason: >-
                    Cargo revertido el 2026-05-14. Cliente notificado por
                    correo electrónico el mismo día.
      responses:
        "200":
          description: Status updated.
          headers:
            traceparent: { $ref: "#/components/headers/TraceparentResponse" }
            ETag: { $ref: "#/components/headers/ETagResponse" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "400": { $ref: "#/components/responses/Problem400" }
        "401": { $ref: "#/components/responses/Problem401" }
        "403": { $ref: "#/components/responses/Problem403" }
        "404": { $ref: "#/components/responses/Problem404" }
        "409": { $ref: "#/components/responses/Problem409" }
        "412": { $ref: "#/components/responses/Problem412" }
        "422": { $ref: "#/components/responses/Problem422" }
        "428": { $ref: "#/components/responses/Problem428" }
        "500": { $ref: "#/components/responses/Problem500" }
        "503": { $ref: "#/components/responses/Problem503" }
...
components:
  securitySchemes:
    mtls:
      type: mutualTLS
      description: |
        Institutions present a client certificate issued by the SBS CA. mTLS
        termination happens at the edge; per-call validation is enforced by
        the API service. Implementation details land in Prompt 7.
    oauth2_client_credentials:
      type: oauth2
      description: |
        OAuth 2.0 client_credentials flow scoped per institution per
        ADR 0032. Tokens are cert-bound via the RFC 8705
        `cnf.x5t#S256` confirmation claim — the token is unusable from
        a different mTLS connection. See [api/openapi/error-catalog.md]
        for the stable error codes returned on token/scope failures.
      flows:
        clientCredentials:
          tokenUrl: /v1/oauth/token
          scopes:
            complaints:write: Submit complaints and patch their resolution status
            complaints:read: List and retrieve complaints
            batch:upload: Submit batch manifests and read batch status / results
            status:read: Read institution operational status (status endpoints)
    hmac_signature:
      type: apiKey
      name: X-SBS-Signature
      in: header
      description: |
        HMAC-SHA256 signature over a canonical request envelope (method, path,
        body hash, timestamp, idempotency key). Detached from the OAuth bearer
        token because both layers serve different purposes: the bearer asserts
        institutional identity at the edge, the HMAC binds a specific request
        body. Implementation details land in Prompt 7.
  headers:
    TraceparentResponse:
      description: |
        W3C traceparent identifier for the request. Echoed verbatim from the
        request when present and well-formed; otherwise generated by the
        server. Present on every response (success and error) so clients can
        correlate logs and traces across the request path.
      schema:
        type: string
        pattern: "^[0-9a-f]{2}-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$"
...
  parameters:
    IdempotencyKeyHeader:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, format: uuid }
      description: |
        UUID v4. Required on every POST and PATCH. Replays of the same key
        within 24 hours return the original response (a cached representation),
        not a fresh execution. Keys are scoped per institution.
    TraceparentHeader:
      in: header
      name: traceparent
      required: false
      schema:
        type: string
        pattern: "^[0-9a-f]{2}-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$"
      description: |
        W3C traceparent identifier. If absent, the server generates one and
        returns it in the response. Every operation echoes the resulting
        traceparent in the response `traceparent` header.
...
  schemas:
...
    ComplaintSubmission:
      type: object
      additionalProperties: false
      required: [complaint]
      properties:
        complaint:
          $ref: "#/components/schemas/Complaint"
        client_submission_id:
          type: [string, "null"]
          minLength: 8
          maxLength: 64
          description: Optional institution-side correlation id.
...
    ResolutionStatus:
      type: string
      enum: [pendiente, atendido, anulado]
      description: Anexo 1-A #22 — Estado.
  examples:
...
    ValidComplaintSubmission:
      summary: A valid complaint submission
      value:
        complaint:
          complaint_id: BCO-2026-000017
          institution_id: SBS-001234
          received_date: 2026-05-12
          complainant_doc_type: DNI
          product_category: TARJETA_CREDITO
          channel: APP_MOVIL
          motivo_code: COBRO_INDEBIDO
          severity: HIGH
          description_text: >-
            Se realizó un cargo no autorizado por S/ 245.00 en la tarjeta
            de crédito el 10 de mayo.
          description_language: es
          complainant_age_range: "35_44"
          complainant_district: "150100"
          submission_method: APP_MOVIL
          original_reference_id: null
          resolution_status: pendiente
        client_submission_id: sub-2026-05-12-abc123
...